+++ NEWS.md
@@ -1,5 +1,7 @@
-# [wasserportal 0.3.1](https://github.com/KWB-R/wasserportal/releases/tag/v0.3.1) <small>2024-04-04</small>
+# [wasserportal 0.4.0](https://github.com/KWB-R/wasserportal/releases/tag/v0.4.0) <small>2024-04-05</small>
 
+* New feature: add support for downloading all available surface water quality 
+data for one or multiple monitoring stations. For details see `get_surfacewater_qualities()` 
 * Bugfix for groundwater level and quality due to new Wasserportal API 
 * Add project [AD4GD](https://www.kompetenz-wasser.de/de/forschung/projekte/ad4gd) 
 as funder